Hot fix when using variant mapping with bgen files
I changed the logic flow to first check whitelist then do variant mapping, this is because if we are using a variant mapping file to map between the genotype snps and model snps the varid gets updated to varid_ when there is a match and varid takes a new id which doesn't match with ids in the whitelist. With this update we ensure that change does not affect the selection at whitelisting step.